Создал data.js & util.js. Внёс измнения в main.js
